PhyNEST

Documentation for a Julia package PhyNEST: Phylogenetic Network Estimation using SiTe patterns.

Tutorials

A step-by-step wiki tutorial is available, that has been done for the 2023 Botany workshop.

Reference

Sungsik Kong, David Swofford, and Laura Kubatko (2022) Inference of Phylogenetic Networks from Sequence Data using Composite Likelihood. Preprint available online on BioRxiv at https://doi.org/10.1101/2022.11.14.516468.

Getting help

Please use google group to report bugs or post questions and/or suggestions.

Functions

PhyNEST.greet — Function

greet()

Displays a simple greet message with citation information. No input argument is needed.

greet()

PhyNEST.readPhylip — Function

readPhylip(inputfile::AbstractString; 
            writecsv=false::Bool,
            csvname=""::AbstractString,
            showProgress=true::Bool,
            tdigits=3::Integer,
            checkpoint=false::Bool)

Import, read, and parse the input phylip file. File name must be specified as string. By default, progress bar is displayed, and .csv and .ckp files are NOT produced. See optional arguments below and modify if needed.

Mandatory argument

inputfile Name of the phylip file as a string

Optional arguments

writecsv (default=false) A boolean arguement to allow writing site pattern frequencies in .csv file

csvname (default=sitePatternCounts_inputfile.csv) A string that will be name of the .csv file

showProgress (default=true) A boolean argument for visualizing the progress of alignment parsing

tdigits (default=3) Number of decimal points to record timme taken to parse the file in seconds

checkpoint (default=false) A boolean argument to store the Phylip object as a .ckp file. (Warning: .ckp file can be large)

PhyNEST.readPhylipFile — Function

readPhylipFile(inputfile::AbstractString,
                writecsv::Bool,
                csvname::AbstractString,
                showProgress::Bool)

Exceuted while running the function readPhylip. Fills in the attributes in the Phylip object, except for the Phylip.time.

PhyNEST.PhylipFileInfo — Function

PhylipFileInfo(inputfile::AbstractString, 
                p::Phylip, 
                showProgress::Bool)

Function that shortens?summarizes? the sequence alignment into two matrices:

The one with unique sites and

Another with how many times each column in the previous matrix occurs throughout the alignment.

Types

PhyNEST.Phylip — Type

Phylip

Subtype of abstract INPUT type with the following attributes:

filename Name of the input phylip alignment file

time Time taken for parsing the input in seconds

numtaxa Number of taxa given at the first line of the input file

seqleng Sequence length given at the first line of the input file

nametaxa Sequence names given in the input file

counttaxa Unique integer identifier given to each individual in the order of appearance in the input file

allquartet All combinations of quartets for counttaxa

index Just convert allquartet elements into arbitrary index numbers

spcounts Arrays of 15 site pattern frequencies for each quartet in allquartet

Index

PhyNEST.Phylip

PhyNEST.PhylipFileInfo

PhyNEST.greet

PhyNEST.readPhylip

PhyNEST.readPhylipFile