PhyNEST

Documentation for a Julia package PhyNEST: Phylogenetic Network Estimation using SiTe patterns.

Tutorials

A step-by-step wiki tutorial is available, that has been done for the 2023 Botany workshop.

Reference

Sungsik Kong, David Swofford, and Laura Kubatko (2022) Inference of Phylogenetic Networks from Sequence Data using Composite Likelihood. Preprint available online on BioRxiv at https://doi.org/10.1101/2022.11.14.516468.

Getting help

Please use google group to report bugs or post questions and/or suggestions.

Functions

PhyNEST.greet — Function

greet()

Displays a greeting with citation information. No argument needed.

Example

julia> greet()
Thank you for using PhyNEST: Phylogenetic Network Estimation using SiTe patterns.
Please report bugs or make suggestions to https://groups.google.com/g/phynest-users.
If you conduct an analysis using PhyNEST, please cite:
Sungsik Kong, David Swofford, and Laura Kubatko (2022) Inference of Phylogenetic Networks from Sequence Data using Composite Likelihood.
Preprint available online on BioRxiv at https://doi.org/10.1101/2022.11.14.516468.

PhyNEST.readPhylip — Function

readPhylip(inputfile::AbstractString; 
            writecsv=false::Bool,
            csvname=""::AbstractString,
            showProgress=true::Bool,
            tdigits=3::Integer,
            checkpoint=false::Bool)

Import, read, and parse the input phylip file. File name must be specified as string. By default, progress bar is displayed, and .csv and .ckp files are NOT produced. See optional arguments below and modify if needed.

Mandatory argument

inputfile Name of the phylip file as a string

Optional arguments

writecsv (default=false) A boolean arguement to allow writing site pattern frequencies in .csv file

csvname (default=sitePatternCounts_inputfile.csv) A string that will be name of the .csv file

showProgress (default=true) A boolean argument for visualizing the progress of alignment parsing

tdigits (default=3) Number of decimal points to record timme taken to parse the file in seconds

checkpoint (default=false) A boolean argument to store the Phylip object as a .ckp file. (Warning: .ckp file can be large)

PhyNEST.readPhylipFile — Function

readPhylipFile(inputfile::AbstractString,
                writecsv::Bool,
                csvname::AbstractString,
                showProgress::Bool)

Exceuted while running the function readPhylip. Fills in the attributes in the Phylip object, except for the Phylip.time.

PhyNEST.PhylipFileInfo — Function

PhylipFileInfo(inputfile::AbstractString, 
                p::Phylip, 
                showProgress::Bool)

Function that shortens?summarizes? the sequence alignment into two matrices:

The one with unique sites and

Another with how many times each column in the previous matrix occurs throughout the alignment.

PhyNEST.getUniqueQuartets — Function

getUniqueQuartets(p::Phylip)

Getting some combinations for four sequences (quartets). This will result in n! quartets where n is the number of sequences. I believe a better way to do this exists.

PhyNEST.sitePatternCounts — Function

sitePatternCounts(p::Phylip,
                ppbase::Array,
                counts::Array)

Computes observed site pattern frequencies from UniqueBase and BaseCounts obtained from the function PhylipFileInfo.

PhyNEST.spRearrange — Function

spRearrange(p::Phylip)

Rearranges the elemetns of each unique quartet obtained using getUniqueQuartets, and also suffles the site pattern frequencies accordingly. This prevents 24 redundant computations, saving computation time.

PhyNEST.show_sp — Function

show_sp(p::Phylip)

Pretty name for displaying all quartet and the corresponding site pattern frequencies on screen in the DataFrame format. It may result in excessively long table when there are many sequences in the input file.

Input

p Phylip object [mandatory]

PhyNEST.sitePatternsToDF — Function

sitePatternsToDF(p::Phylip)

Extracts the quartet and site pattern information from the Phyliip object, and reorganizes them in the DataFrame format. This function does all hard work work for show_sp.

PhyNEST.write_sp — Function

write_sp(p::Phylip; 
        csvname="PhyNEST_sp"::AbstractString)

Pretty name for exporting all quartet and the corresponding site pattern frequencies in a .csv format.

Input

p Phylip object [mandatory] csvname Filename for the .csv output can be given, otherwise we use PhyNEST_sp.csv by default.

PhyNEST.storeCheckPoint — Function

storeCheckPoint(p::Phylip)

Stores the phylip object into a .ckp file so a user does not have to repeat the phylip parsing again if working with the same dataset next time. By default it creates a .ckp file that has the same filename as the input phylip file. File size can get pretty large...

Input

p Phylip object [mandatory]

PhyNEST.readCheckPoint — Function

readCheckPoint(ckpfile::AbstractString)

Reads in .ckp file and creates a filled in phylip object.

Input

ckpfile Name of the checkpoint file

PhyNEST.GetTrueProbsSymm — Function

GetTrueProbsSymm(myt1::Float64,
                myt2::Float64,
                myt3::Float64,
                theta::Float64)
GetTrueProbsSymm(myt1::Float64,
                myt2::Float64,
                myt3::Float64,
                theta::Float64,
                alpha::Float64)

Computes true site pattern probabilities for the symmetric quartet tree: ((1,2),(3,4));. Three speciation times (node ages) in coalescent unit and theta must be provided; alpha is assumed to be 4/3 if unspecified. See the manuscript and/or Chifman and Kubatko (2015)[10.1016/j.jtbi.2015.03.006] for more information.

Input

-myt1 Speciation time for the common ancestor of species 1 and 2 in coalescent unit -myt2 Speciation time for the common ancestor of species 3 and 4 in coalescent unit -myt3 Root node age in coalescent unit -theta Effective population size parameter -alpha (dafault=4/3)

PhyNEST.GetTrueProbsAsymm — Function

GetTrueProbsAsymm(myt1::Float64,
                myt2::Float64,
                myt3::Float64,
                theta::Float64)
GetTrueProbsAsymm(myt1::Float64,
                myt2::Float64,
                myt3::Float64,
                theta::Float64,
                alpha::Float64)

Computes true site pattern probabilities for the asymmetric quartet tree: (1,(2,(3,4)));. Three speciation times (node ages) in coalescent unit and theta must be provided; alpha is assumed to be 4/3 if unspecified. See the manuscript and/or Chifman and Kubatko (2015)[10.1016/j.jtbi.2015.03.006] for more information.

Input

-myt1 Speciation time for the common ancestor of species 3 and 4 in coalescent unit -myt2 Speciation time for the common ancestor of species 2 and (3,4) in coalescent unit -myt3 Root node age in coalescent unit -theta Effective population size parameter -alpha (dafault=4/3)

PhyNEST.Dstat — Function

Dstat(outgroup::String, p::Phylip;
    pval=0.05::Float64, 
    displayall=false::Bool)

Conducts Patterson's D-statistic test. The result prints site pattern frequencies ABAB and ABBA used to compute the D-statistic, Z-score, and the p-value for each quartet tested. Significance is marked with an asterisk. Function showall(df) can be subsequently used to show all rows.

Mandatory arguments

outgroup Name of the outgroup taxa

p The Phylip object

Optional arguments

pval (default=0.05) Alpha level for significance

display_all (default=false) If set as true, the function print test results for every quartet. By default, it only prints those quartets where signficance was found.

Example

julia> p=readPhylip("sample_n4h1.txt")
julia> df=Dstat("4",p)
Tip: if neccessary, use showall(df) function to see all the rows.
2×10 DataFrame
 Row │ outgroup  taxa1   taxa2   taxa3   ABAB   ABBA   Dstat     Zscore   pvalue   significance
     │ String    String  String  String  Int64  Int64  Float64   Float64  Float64  String
─────┼──────────────────────────────────────────────────────────────────────────────────────────
   1 │ 4         3       2       1        1427   7852  0.692424  66.6995      0.0  *
   2 │ 4         1       2       3        1427   7836  0.691892  66.5908      0.0  *

julia> df=Dstat("4",p,display_all=true)
Tip: if neccessary, use showall(df) function to see all the rows.
6×10 DataFrame
Row │ outgroup  taxa1   taxa2   taxa3   ABAB   ABBA   Dstat        Zscore      pvalue    significance
    │ String    String  String  String  Int64  Int64  Float64      Float64     Float64   String
────┼─────────────────────────────────────────────────────────────────────────────────────────────────
  1 │ 4         3       1       2        7852   1427  -0.692424    -66.6995    1.0
  2 │ 4         3       2       1        1427   7852   0.692424     66.6995    0.0       *
  3 │ 4         1       3       2        7836   1427  -0.691892    -66.5908    1.0
  4 │ 4         1       2       3        1427   7836   0.691892     66.5908    0.0       *
  5 │ 4         2       3       1        7836   7852   0.00101989    0.127743  0.449176
  6 │ 4         2       1       3        7852   7836  -0.00101989   -0.127743  0.550824

PhyNEST.showall — Function

showall(df::DataFrame)

Print all rows of the DataFrame object.

Types

PhyNEST.Phylip — Type

Phylip

Subtype of abstract INPUT type with the following attributes:

filename Name of the input phylip alignment file

time Time taken for parsing the input in seconds

numtaxa Number of taxa given at the first line of the input file

seqleng Sequence length given at the first line of the input file

nametaxa Sequence names given in the input file

counttaxa Unique integer identifier given to each individual in the order of appearance in the input file

allquartet All combinations of quartets for counttaxa

index Just convert allquartet elements into arbitrary index numbers

spcounts Arrays of 15 site pattern frequencies for each quartet in allquartet

Index

PhyNEST.Phylip

PhyNEST.Dstat

PhyNEST.GetTrueProbsAsymm

PhyNEST.GetTrueProbsSymm

PhyNEST.PhylipFileInfo

PhyNEST.getUniqueQuartets

PhyNEST.greet

PhyNEST.readCheckPoint

PhyNEST.readPhylip

PhyNEST.readPhylipFile

PhyNEST.show_sp

PhyNEST.showall

PhyNEST.sitePatternCounts

PhyNEST.sitePatternsToDF

PhyNEST.spRearrange

PhyNEST.storeCheckPoint

PhyNEST.write_sp