PhyNE.jl

Documentation for a julia package PhyNE.jl: Phylogenetic Network Estimation using composite likelihood directly from the sequence data using site pattern frequency.

References

TBA

Getting help

Please use google group to report bugs or make suggestions.

Manual

Installation

Julia installation

Installation of the package PhyNE

Test example

Input

Parsing DNA alignment data

Reading .ckp file

Exporting as .csv file

Quartet

Reading a network

Extract quartet(s)

True site pattern probabilities for a quartet

Simulate true site pattern frequencies

Method-of-moment estimator of branch lengths

Estimating theta

Network estimation

Optimizing parameters for a given network

Network analysis

Visualizing network

Others

HyDe

The D-statistics

Functions

PhyNE.readPhylipFile! — Function

readPhylipFile!(inputfile::AbstractString, writecsv::Bool)
readPhylipFile!(inputfile::AbstractString)

Takes in, read, and parse the input phylip file. File name is specified as a string. If writecsv=true, list of site pattern frequencies from the input phylip for all permutations of four taxa is saved as a .csv file in the working directory. The writecsv boolean is false by default. The function will perform assuming writecsv=false if not specified.

Input

inputfile Name of phylip file as a AbstractString
writecsv Boolean to write site pattern frequencies in CSV file

PhyNE.readCheckPoint — Function

readCheckPoint(ckpfile::AbstractString)

Reads in .ckp file. gegeege

PhyNE.writeSitePatternCounts — Function

writeSitePatternCounts(p::Phylip,write::Bool,inputfile::AbstractString)
writeSitePatternCounts(p::Phylip,inputfile::AbstractString)
writeSitePatternCounts(inputfile::AbstractString)
writeSitePatternCounts(p::Phylip)

Writes the observed site pattern frequencies from the phylip file as .csv file in the working directory and name it as sitePatternCounts_inputfile.csv. Although it requires an attribute inputfile, it does not actually read the file as long as the Phylip object in memory is provided. (i.e., any AbstractString can be provided here that is used as a suffix of the ) However, if only the input file name is provided, it runs readPhylipFile! first then write csv. file.

Input

p A Type Phylip object
write Boolean variable to export site pattern frequences in .csv
inputfile Name of phylip file as a AbstractString

Missing docstring.

Missing docstring for mrca. Check Documenter's build log for details.

PhyNE.nature — Function

nature(Node::Int64,net::HybridNetwork)

Function to determine the nature (root, internal tree node, leaf, hybrid node) of node given the node number and Hybridnetwork.

Input

Node An integer node number

net A tree/network in HybridNetwork

PhyNE.extractNQuartets — Function

extractNQuartets(net::HybridNetwork,p::Phylip)

gege

Missing docstring.

Missing docstring for printQuarts. Check Documenter's build log for details.

PhyNE.GetTrueProbsSymm — Function

GetTrueProbsSymm(myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64,alpha::Float64)

gege

PhyNE.GetTrueProbsAsymm — Function

GetTrueProbsAsymm(myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64,alpha::Float64)

gege

PhyNE.GetTrueProbsNetTypes — Function

GetTrueProbsNetTypes(type::Integer,myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64,alpha::Float64)

gege

PhyNE.simspcounts — Function

simspcounts(type::Integer,myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64,alpha::Float64,n::Integer)

This function simulates the 15 site pattern probabilities for a quartet.

- `type`   specify the type of a quartet. 
- `myt1`   Branch length
- `myt2`   Branch length
- `myt3`   Branch length from the root
- `theta`  Theta value (default=0.01)
- `alpha`  Alpha value (default=4/3)
- `length` Sequence lengths in integers (default=1000000)

Missing docstring.

Missing docstring for momentEstimat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for startTheta. Check Documenter's build log for details.

PhyNE.Optimization — Function

Optimization

hehe

PhyNE.hillClimb — Function

function hillClimb(startT::HybridNetwork,p::Phylip,hmax::Integer,outgroup::String,maxcount::Integer,
nfail::Integer,NumIter::Integer,paramprint::Bool,logfile::IO,display::Bool)

Executes Hill-Climbing Algorithm.

Missing docstring.

Missing docstring for maxburnin. Check Documenter's build log for details.

Missing docstring.

Missing docstring for simAnneal. Check Documenter's build log for details.

PhyNE.PhyNe! — Function

PhyNe!(startT::HybridNetwork,inputFile::Phylip,outgroup::String)

Estimates the network or tree to fit observed site pattern frequencies stored in a Phylip object, using composite likelihood. A level-1 network is assumed. The search begins from topolgoy startT, which can be a tree or a network (with less than hmax reticulation nodes). Usually, startT is expected to be a tree estiamted from data, but can also be randomly generated. The topology is rooted using the outgroup provided. This must be identical to the outgroup sequence provided in the sequence alignment. By default, PhyNe! will search the network space using simulated annealing assuming hmax=1.

There are many optional arguments (values in parenthesis are defaults):

hillclimbing(=false) Select hill climbing search

hmax(=1) Maximum number of reticulations in the final network

nruns(=10) Number of independent runs

nniStartT=false

maxcount=1000 Number of steps to terminate the search

nfail=75 Number of consecutive failures to terminate the search

Simulated annealing stuff:

burninn=25

k=10

ProbabilityOfNotSelectingNode=9.5e-45

cons=0.9

alph=0.8

Optimization stuff:

NumIter=1000

paramprint=false

Miscellaneous:

timestamp=false

filename=""

display=falses

Missing docstring.

Missing docstring for Dstat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HyDe. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LRT. Check Documenter's build log for details.

Types

PhyNE.Phylip — Type

Phylip

Subtype of abstract INPUT type. Sequence alignment with the following attributes:

filename The actual filename of the input sequence alignment

time Time it took to intake the filename in seconds

numtaxa The number of taxa appears on the first line of phylip

seqleng The sequence length appears on the first line of phylip

nametaxa Names of each individual in the order of appearance in phylip

counttaxa Numbering each individual in the order of appearance in phylip

allquartet All combinations of quartets using nametaxa

index Just convert allquartet elements into arbitrary index numbers

spcounts Arrays of 15 site pattern frequencies

PhyNE.nquartets — Type

nquartets

Subtype of abstract NQuartet type. Stores relevant information of a quartet extracted from a topology with the following attributes:

number List of individuals quartets in the order of i,j,k,l

tquartet List of quartets in the order of i,j,k,l

indexNET Index number for each quartet in the order appears in quartnet.

mspcountsNET We can directly use this counts for likelihood calculation.

mrca List of common ancesters of two taxa in the order of i and j (ij),ik,il,jk,jl,and kl

ntau Unique number of the taus in the tau used in branchlengths

branchlength Branch length from the mrca to tau=0 in the order of i and j (ij),ik,il,jk,jl,and kl

momestlength Branch length identified for each quartet given the mspcounts using moment estimator - Will get filled later because theta is required

mombl Branch length that is averaged for the entire tree/network - Will get filled later because theta is required

symtype Type of each quartet. It can be either symmetric or asymmetric. Asymmetric (type 0) can have four possible topologies and we define type 1 as (i,((j,k),l)); type 2 as ((i,(j,k)),l); type 3 as (i,(j,(k,l))); and type 4 as (((i,j),k),l).

logLik It's just a likelihood for the quartet

PhyNE.Nquartets — Type

Nquartets

Subtype of abstract NQuartet type. Stores relevant information of a tree extracted from a topology with the following attributes:

leafname Name of the leaves as appears on the newick

leafnumber Assigned number for each leaf when reading in the newick

gamma Gamma as appears on the newick, although irrelevant for our computation

nquartet An array of all quartets in the given topology

Index

PhyNE.Nquartets

PhyNE.Phylip

PhyNE.nquartets

PhyNE.GetTrueProbsAsymm

PhyNE.GetTrueProbsNetTypes

PhyNE.GetTrueProbsSymm

PhyNE.Optimization

PhyNE.PhyNe!

PhyNE.extractNQuartets

PhyNE.hillClimb

PhyNE.nature

PhyNE.readCheckPoint

PhyNE.readPhylipFile!

PhyNE.simspcounts

PhyNE.writeSitePatternCounts