PhyNEST.jl

Documentation for a Julia package PhyNEST.jl: Phylogenetic Network Estimation using SiTe patterns.

References

Sungsik Kong, David Swofford, and Laura Kubatko (2022) Inference of Phylogenetic Networks from Sequence Data using Composite Likelihood. Preprint available online on BioRxiv at https://doi.org/10.1101/2022.11.14.516468.

Getting help

Please use google group to report bugs or post questions and/or suggestions.

Manual

Installation

Julia installation

Installation of the package PhyNEST

Test example

Input

Parsing DNA alignment data

Reading .ckp file

Network estimation

Getting ready for the network estimation

Hill climbing

Simulated annealing

Functions

PhyNEST.greet — Function

greet()

Displays a simple greet message with citation information. No input argument is needed.

PhyNEST.readPhylip — Function

readPhylip(inputfile::AbstractString)

Import, read, and parse the input phylip file. File name must be specified as string.
By default, .csv and .ckp files are NOT produced. See below for the optional arguments.\

To save the observed quartet site pattern frequencies in a .csv file, set the optional argument
writecsv=true. The .csv file will be stored in the working directory, with a filename
"sitePatternCounts_inputfile.csv". If a user would like to modify the filename, it can be done
by providing preferred name using the optional argument csvname="preferred_file_name.csv".

Input

Mandatory

inputfile Name of the phylip file as a String [mandatory]\

Optional

writecsv (default=false) A Boolean arguement to write site pattern frequencies in CSV file

csvname (default=sitePatternCounts_inputfile.csv) A string that will be name of the .csv file

showProgress (default=true) A boolean argument for visualizing the progress of alignment parsing process

tdigits (default=3) Number of decimal points to record timme taken to parse the file in seconds

checkpoint (default=false) A boolean argument to store Phylip object as a .ckp file. (Warning: .ckp can be large).

PhyNEST.show_sp — Function

show_sp(p::Phylip)

Pretty name for displaying all quartet and the corresponding site pattern frequencies on screen in the DataFrame format.
It may result in excessively long table when there are many sequences in the input file.

Input

p Phylip object [mandatory]

PhyNEST.write_sp — Function

write_sp(p::Phylip)

Pretty name for exporting all quartet and the corresponding site pattern frequencies in a .csv format.

Input

p Phylip object [mandatory] csvname Filename for the .csv output can be given, otherwise we use PhyNEST_sp.csv by default.

PhyNEST.storeCheckPoint — Function

storeCheckPoint(p::Phylip)

Stores the phylip object into a .ckp file so a user does not have to repeat the phylip parsing again if working with the
same dataset next time. By default it creates a .ckp file that has the same filename as the input phylip file.
File size can get pretty large...

Input

p Phylip object [mandatory]

PhyNEST.readCheckPoint — Function

readCheckPoint(ckpfile::AbstractString)

Reads in .ckp file and creates a filled in phylip object.

Input

ckpfile Name of the checkpoint file

PhyNEST.GetTrueProbsSymm — Function

GetTrueProbsSymm(myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64)
GetTrueProbsSymm(myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64,alpha::Float64)

Computes true site pattern probabilities for the symmetric quartet tree, ((1,2),(3,4));.
Three speciation times (or node ages) in coalescent unit and theta must be provided. Alhpa is not
an essential argument for the function and if not provided, it isssumed to be 4/3 by default.
See manuscript or Chifman and Kubatko (2015)[10.1016/j.jtbi.2015.03.006] for more information.

Input

myt1 Speciation time for the common ancestor of species 1 and 2
myt2 Speciation time for the common ancestor of species 3 and 4
myt3 Root node age
theta Effective population size parameter
alpha 4/3 by default

PhyNEST.GetTrueProbsAsymm — Function

GetTrueProbsAsymm(myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64)
GetTrueProbsAsymm(myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64,alpha::Float64)

Computes true site pattern probabilities for the asymmetric quartet tree, (1,(2,(3,4)));.
Three speciation times (or node ages) in coalescent unit and theta must be provided. Alhpa is not
an essential argument for the function and if not provided, it isssumed to be 4/3 by default.
See manuscript or Chifman and Kubatko (2015)[10.1016/j.jtbi.2015.03.006] for more information.

Input

myt1 Speciation time for the common ancestor of species 3 and 4
myt2 Speciation time for the common ancestor of species 2, 3 and 4
myt3 Root node age
theta Effective population size parameter
alpha 4/3 by default

PhyNEST.GetTrueProbsAsymmTypes — Function

GetTrueProbsAsymmTypes(type::Integer,myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64)
GetTrueProbsAsymmTypes(type::Integer,myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64,alpha::Float64)

Computes true site pattern probabilities for any of the four the asymmetric quartet trees:

Type 1: (i,((j,k),l));

Type 2: ((i,(j,k)),l);

Type 3: (i,(j,(k,l)));

Type 4: (((i,j),k),l).

Three speciation times (or node ages) in coalescent unit and theta must be provided.
Alhpa is not an essential argument for the function and if not provided, it isssumed to be 4/3 by default.
See manuscript or Chifman and Kubatko (2015)[10.1016/j.jtbi.2015.03.006] for more information.

Input

myt1 Speciation time for the common ancestor of species 3 and 4
myt2 Speciation time for the common ancestor of species 2, 3 and 4
myt3 Root node age
theta Effective population size parameter
alpha 4/3 by default

PhyNEST.sim_sp_counts — Function

sim_sp_counts(type::Integer,myt1::Float64,myt2::Float64,myt3::Float64)
sim_sp_counts(type::Integer,myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64,alpha::Float64,length::Integer)

This function simulates the 15 site pattern probabilities for five quartet tree topologies,
one symmetric and four asymmetric quartets:

Type 0: ((i,j),(k,l));

Type 1: (i,((j,k),l));

Type 2: ((i,(j,k)),l);

Type 3: (i,(j,(k,l)));

Type 4: (((i,j),k),l).

Input

type Specify the type of a quartet (use integer)
myt1 Speciation time for the most recent internal tree node. Common ancestor of 1 and 2 in the symmetric case
myt2 Speciation time for the internal tree node closer to the root
myt3 Root node age
theta Effective population size parameter (default=0.01)
alpha Alpha (default=4/3)
length Sequence lengths (default=1000000)

PhyNEST.get_quartets — Function

get_quartets(net::HybridNetwork; round_gamma_digits=5::Int64)
get_quartets(net::HybridNetwork, p::Phylip; round_gamma_digits=5::Int64, default_theta=0.0001::Float64)

Extracts information for every quartet extractable from the topology and store it as a Network object.
When the Phylip object is also provided, we fill more slots including the theta, branch lengths for
eac quartet using the moment estimator, average branch lengths using the estimated branch lengths.

Input

net PhyloNetwork.HybridNetwork object for a topology [mandatory]

p Phylip object

round_gamma_digits Number of digits to show for the inhertiance probability in each quartet

`default_theta Theta is being estimated when both topology and data are given, however,
when the estimated theta is beyond the expected range (i.e., 0.00001 < theta < 0.1)
we use user-specified theta (default=0.0001) to estimate branch lengths using
the moment estimator.

Missing docstring.

Missing docstring for printQuartets. Check Documenter's build log for details.

PhyNEST.get_start_theta — Function

get_start_theta(N::Network; 
                lower_bound=0.00001::Float64,factor=2.0::Float64,tolerance=0.01::Float64)

Using the Network object that contains information on quartets extracted from the topology and
site pattern frequencies from the data, estimates the `reasonable' theta value that can be
used as the starting point for the composite likelihood optimization.
Approximate interval where the true theta may lie is estimated through two procedures:
somewhat loose interval that keeps the branch lengths positive from the moment estimator,
and then further tightens the interval using the golden section seach.

PhyNEST.do_optimization — Function

do_optimization(net::HybridNetwork, p::Phylip;
                    lower_bound=0.00001::Float64,factor=2.0::Float64,tolerance=0.01::Float64,
                    number_of_itera=1000::Int64)

Optimizes the composite likelihood of the network topology given the data.

PhyNEST.hill_climbing — Function

hill_climbing(starting_topology::HybridNetwork,p::Phylip,outgroup::String;
    hmax::Integer,
    maximum_number_of_steps::Integer,
    maximum_number_of_failures::Integer,
    number_of_itera::Integer)

Executes hill climbing algorithm for the network space search.

PhyNEST.simulated_annealing — Function

simulated_annealing(starting_topology::HybridNetwork, p::Phylip, outgroup::String, 
    hmax::Integer, 
    number_of_burn_in::Int64,
    maximum_number_of_steps::Integer,
    maximum_number_of_failures::Integer,
    number_of_itera::Integer,
    k::Integer,
    cons::Float64,
    alph::Float64
    )

Executes simulated algorithm for the network space search.

PhyNEST.phyne! — Function

phyne!(starting_topology::HybridNetwork,p::Phylip,outgroup::String;
    hmax=1::Integer,
    maximum_number_of_steps=1000::Integer,
    maximum_number_of_failures=100::Integer,
    number_of_itera=1000::Integer,
    number_of_runs=5::Integer,
    do_hill_climbing=true::Bool,
    number_of_burn_in=25::Integer,
    k=10::Integer,
    cons=0.5::Float64,
    alph=0.5::Float64
    )

phyne! is fine.

Types

PhyNEST.Phylip — Type

Phylip

Subtype of abstract INPUT type with the following attributes:

filename Name of the input phylip alignment file

time Time taken for parsing the input in seconds

numtaxa Number of taxa given at the first line of the input file

seqleng Sequence length given at the first line of the input file

nametaxa Sequence names given in the input file

counttaxa Unique integer identifier given to each individual in the order of appearance in the input file

allquartet All combinations of quartets for counttaxa

index Just convert allquartet elements into arbitrary index numbers

spcounts Arrays of 15 site pattern frequencies for each quartet in allquartet

PhyNEST.Network — Type

Network

Subtype of abstract Quartet type with the following attributes:

leafname Name of the leaves in the order of appearance in the input newick

leafnumber Assigned number for each leaf when reading in the input newick

gamma Gamma as appears on the newick, although irrelevant for our computation

theta Estimated theta using the site pattern frequencies. If Network object is
created without Phylip, theta is not estimated byt default theta=0.001 is displayed.

quartet An array of all quartets in the given topology

PhyNEST.quartets — Type

quartets

Subtype of abstract Quartet type with the following attributes:

number List of individuals quartets in the order of i,j,k,l
displayed_treenth displayed tree that the quartet was extracted from
quartet List of quartets in the order of i,j,k,l using the leaf numbers in HybridNetwork
tquartet List of quartets in the order of i,j,k,l using the leaf numbers in Phylip
gamma Inhertiance probability information provided in HybridNetwork
mspcountsNET We can directly use this counts for likelihood calculation.
mrca List of common ancesters of two taxa in the order of i and j (ij),ik,il,jk,jl,and kl
ntau Unique number of the taus in the tau used in branchlengths
momestlength Branch length identified for each quartet given the mspcounts using moment estimator
average_mom_est_bl Branch length that is averaged for the entire tree/network - Will get filled later because theta is required
symtype Type of each quartet. It can be either symmetric (type 0) or asymmetric. Asymmetric quartets have four possible topologies:

Type 1: (i,((j,k),l));

Type 2: ((i,(j,k)),l);

Type 3: (i,(j,(k,l)));

Type 4: (((i,j),k),l).

logLik It's just a negative likelihood for the quartet in interest

Index

PhyNEST.Network

PhyNEST.Phylip

PhyNEST.quartets

PhyNEST.GetTrueProbsAsymm

PhyNEST.GetTrueProbsAsymmTypes

PhyNEST.GetTrueProbsSymm

PhyNEST.do_optimization

PhyNEST.get_quartets

PhyNEST.get_start_theta

PhyNEST.greet

PhyNEST.hill_climbing

PhyNEST.phyne!

PhyNEST.readCheckPoint

PhyNEST.readPhylip

PhyNEST.show_sp

PhyNEST.sim_sp_counts

PhyNEST.simulated_annealing

PhyNEST.storeCheckPoint

PhyNEST.write_sp