PhyNEST.jl

Documentation for a julia package PhyNEST.jl: Phylogenetic Network Estimation using SiTe patterns.

References

TBA

Getting help

Please use google group to report bugs or post questions and/or suggestions.

Manual

Installation

Julia installation

Installation of the package PhyNE

Test example

Input

Parsing DNA alignment data

Reading .ckp file

Exporting as .csv file

Network estimation

Simulated Annealing

Hill climbing

Quartet

Reading a network

Extract quartet(s)

True site pattern probabilities for a quartet

Simulate true site pattern frequencies

Method-of-moment estimator of branch lengths

Estimating theta

Functions

PhyNEST.readPhylipFile! — Function

readPhylipFile!(inputfile::AbstractString)
readPhylipFile!(inputfile::AbstractString; writecsv=false::Bool, csvname=""::AbstractString, showProgress=true::Bool)

Takes in, read, and parse the input phylip file. File name must be specified, in string format. There are a number of optional arguments (see below).

Input

inputfile Name of phylip file as a AbstractString
writecsv A Boolean arguemtn to write site pattern frequencies in CSV file (Default=false)
csvname A string that will be name of the .csv file (Default=sitePatternCounts_inputfile.csv)
showProgress A boolean argument for visualizing the progress of alignment parsing process (Default=true)

PhyNEST.readCheckPoint — Function

readCheckPoint(ckpfile::AbstractString)

Reads in .ckp file that is created every time function readPhylipFile! is complete.

Input

ckpfile Name of checkpoint file as a AbstractString

PhyNEST.writeSitePatternCounts — Function

writeSitePatternCounts(phylip::Phylip,csvname::AbstractString)

Exports the site pattern frequencies parsed frin the sequence alignment in a .csv file. The output is stored in the worksing directory unless specified, and named as sitePatternCounts_csvname.csv.

Input

phylip A Phylip object that is obtained using the function readPhylipFile!()
csvname A string that specifies the name of the .csv file

PhyNEST.extractNQuartets — Function

extractNQuartets(net::HybridNetwork, phylip::Phylip)

Extracts quartet information from a topology. Also stores the observed site patter frequency information for each quartet.

Input

net HybridNetwork object obtained using the function readTopology
phylip Phylip object obtained using the function readPhylipFile!

PhyNEST.extractQuartets — Function

extractQuartets(net::HybridNetwork)

Extracts quartet information from a topology.

Input

net HybridNetwork object obtained using the function readTopology

PhyNEST.printQuartets — Function

printQuartets(net::HybridNetwork)

Visualizes quartet information extracted from a topology in the form of DataFrame.

Input

net HybridNetwork object obtained using the function readTopology

PhyNEST.TrueSitePatternSymm — Function

TrueSitePatternSymm(myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64)
TrueSitePatternSymm(myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64,alpha::Float64)

Computes true site pattern probabilities for the symmetric quartet tree, ((1,2),(3,4));. Three speciation times (or node ages) in coalescent unit and theta must be provided. Alhpa is not an essential argument for the function and if not provided, it isssumed to be 4/3 by default. See manuscript or Chifman and Kubatko (2015)[10.1016/j.jtbi.2015.03.006] for more information.

Input

myt1 Speciation time for the common ancestor of species 1 and 2
myt2 Speciation time for the common ancestor of species 3 and 4
myt3 Root age
theta Effective population size parameter
alpha 4/3 by default

PhyNEST.TrueSitePatternAsymm — Function

TrueSitePatternAsymm(myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64)
TrueSitePatternAsymm(myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64,alpha::Float64)

Computes true site pattern probabilities for the asymmetric quartet tree, (1,(2,(3,4)));. Three speciation times (or node ages) in coalescent unit and theta must be provided. Alhpa is not an essential argument for the function and if not provided, it isssumed to be 4/3 by default. See manuscript or Chifman and Kubatko (2015)[10.1016/j.jtbi.2015.03.006] for more information.

Input

myt1 Speciation time for the common ancestor of species 3 and 4
myt2 Speciation time for the common ancestor of species 2, 3 and 4
myt3 Root age
theta Effective population size parameter
alpha 4/3 by default

PhyNEST.simspcounts — Function

simspcounts(type::Integer,myt1::Float64,myt2::Float64,myt3::Float64)
simspcounts(type::Integer,myt1::Float64,myt2::Float64,myt3::Float64,theta::Float64,alpha::Float64,length::Integer)

This function simulates the 15 site pattern probabilities for five quartet tree topologies, one symmetric and four asymmetric quartets, which we categorize them into five types:

Type 0: ((1,2),(3,4));

Type 1: (1,((2,3),4));

Type 2: ((1,(2,3)),4);

Type 3: (1,(2,(3,4)));

Type 4: (((1,2),3),4);

Input

type Specify the type of a quartet (use integer)
myt1 Speciation time for the most recent internal tree node. Common ancestor of 1 and 2 in the symmetric case
myt2 Speciation time for the internal tree node closer to the root
myt3 Root age
theta Effective population size parameter (default=0.01)
alpha Alpha (default=4/3)
length Sequence lengths (default=1000000)

PhyNEST.PhyNE! — Function

PhyNE!(startT::HybridNetwork,inputFile::Phylip,outgroup::String)

Estimates the network or tree to fit observed site pattern frequencies stored in a Phylip object, using composite likelihood. A level-1 network is assumed. The search begins from topolgoy startT, which can be a tree or a network (with less than hmax reticulation nodes). Usually, startT is expected to be a tree estiamted from data, but can also be randomly generated. The topology is rooted using the outgroup provided. This must be identical to the outgroup sequence provided in the sequence alignment. By default, PhyNE! will search the network space using simulated annealing assuming hmax=1.

There are many optional arguments (values in parenthesis are defaults):

hillclimbing(=false) Select hill climbing search

hmax(=1) Maximum number of reticulations in the final network

nruns(=10) Number of independent runs

nniStartT=false

maxcount=1000 Number of steps to terminate the search

nfail=75 Number of consecutive failures to terminate the search

Simulated annealing stuff:

burninn=25

k=10

ProbabilityOfNotSelectingNode=9.5e-45

cons=0.9

alph=0.8

Optimization stuff:

NumIter=1000

paramprint=false

Miscellaneous:

timestamp=false

filename=""

display=falses

Types

PhyNEST.Phylip — Type

Phylip

Subtype of abstract INPUT type. Sequence alignment with the following attributes:

filename The actual filename of the input sequence alignment

time Time it took to intake the filename in seconds

numtaxa The number of taxa appears on the first line of phylip

seqleng The sequence length appears on the first line of phylip

nametaxa Names of each individual in the order of appearance in phylip

counttaxa Numbering each individual in the order of appearance in phylip

allquartet All combinations of quartets using nametaxa

index Just convert allquartet elements into arbitrary index numbers

spcounts Arrays of 15 site pattern frequencies

PhyNEST.nquartets — Type

nquartets

Subtype of abstract NQuartet type. Stores relevant information of a quartet extracted from a topology with the following attributes:

number List of individuals quartets in the order of i,j,k,l

tquartet List of quartets in the order of i,j,k,l

indexNET Index number for each quartet in the order appears in quartnet.

mspcountsNET We can directly use this counts for likelihood calculation.

mrca List of common ancesters of two taxa in the order of i and j (ij),ik,il,jk,jl,and kl

ntau Unique number of the taus in the tau used in branchlengths

branchlength Branch length from the mrca to tau=0 in the order of i and j (ij),ik,il,jk,jl,and kl

momestlength Branch length identified for each quartet given the mspcounts using moment estimator - Will get filled later because theta is required

mombl Branch length that is averaged for the entire tree/network - Will get filled later because theta is required

symtype Type of each quartet. It can be either symmetric or asymmetric. Asymmetric (type 0) can have four possible topologies and we define type 1 as (i,((j,k),l)); type 2 as ((i,(j,k)),l); type 3 as (i,(j,(k,l))); and type 4 as (((i,j),k),l).

logLik It's just a likelihood for the quartet

PhyNEST.Nquartets — Type

Nquartets

Subtype of abstract NQuartet type. Stores relevant information of a tree extracted from a topology with the following attributes:

leafname Name of the leaves as appears on the newick

leafnumber Assigned number for each leaf when reading in the newick

gamma Gamma as appears on the newick, although irrelevant for our computation

nquartet An array of all quartets in the given topology

Index

PhyNEST.Nquartets

PhyNEST.Phylip

PhyNEST.nquartets

PhyNEST.PhyNE!

PhyNEST.TrueSitePatternAsymm

PhyNEST.TrueSitePatternSymm

PhyNEST.extractNQuartets

PhyNEST.extractQuartets

PhyNEST.printQuartets

PhyNEST.readCheckPoint

PhyNEST.readPhylipFile!

PhyNEST.simspcounts

PhyNEST.writeSitePatternCounts